Use original list when creating a sublist of a sublist
When creating a sublist from a sublist the implementation of subList
from AbstractList is used. This implementation treats the parent
sublist as a list in its own right, and this leads to the child
sublist holding a reference to the parent sublist, when holding a
reference to the original list would suffice. This prevents the parent
sublist from being ellegible for garbage collection as long as the
child sublist is needed, and can lead to StackOverflowError or
OutOfMemoryError errors if multiple calls to subList are chained.
Moreover, it can also make accessing elements of the sublist slower,
as each sublist must call the get function from the sublist it
originated from to access the element.

With this change, when calling subList from a sublist, we return a
sublist holding a reference to the original list instead of to the
parent sublist, which frees the parent sublist for garbage collection
and allows the child sublist to access elements by calling directly the
get function of the original list.

^KT-50081 Fixed
^KT-46103 Fixed