[IR] KT-57181: Properly map type parameters when copying default values
First, default parameters par se are not problematic:
the `functionFromOtherModule.kt` test passes as is.

Here's what was happening in `withTypeParameter.kt`: previously,
when calling `findDeserializedDeclarationForSymbol`
for `deserialized class Any` and trying to do
`deserializeAllReachableTopLevels()`
we are deserializing the actual `topLevel` version
(`box.multiplatform.k2.defaultArgument
.withTypeParameter/topLevel|8497457487728652438[0]`), in particular
its value parameters, and ultimately the default
value of the last parameter, which has a functional
type `(T) -> Int`. But `T` references the
type parameter from the expect function.
This happens because when we copy the default
value of the value parameter, we fail to replace the
type parameter symbol, because of the missing mapping.

^KT-57181 Fixed