[Analysis API FIR] Improve the performance of RemoveRedundantQualifierNameInspection in the K2 IDE
Previously, for every qualified statement, we reanalyzed the whole
containing declaration to get FirTowerContextProvider for it.
This operation is heavyweight, so we cache the results on a per-file basis.
Now we will execute `LowLevelFirApiFacadeForResolveOnAir.onAirGetTowerContextProvider.`
for every declaration only a single time.

Cached FirTowerContextProvider consumes a lot of RAM,
so we need to evict them when they are not needed.
The caffeine library is used for a cache as it allows easy eviction of unneeded elements.

On the Fe10GenerateTests.kt file from the KTIJ-25125,
the RemoveRedundantQualifierNameInspection now takes about 70 samples,
before it was about 800 samples.

^KTIJ-25125 fixed