KT-41422 Improve performance of functions calling collectionSizeOrDefault
Different extension functions declared for iterables use
`Iterable::collectionSizeOrDefault` to find out desired size of
a collection they will return. `collectionSizeOrDefault` checks if its
argument is an instance of a class implementing `Collection` interface.
That check is unnecessary if we know exactly that a function was
invoked on an instance of `Collection`.
To exploit that optimization a new set of extension functions where
defined for `Collection`:
- all, any, none;
- count;
- associate, associateBy, associateWith;
- map, mapIndexed;
- runningFold, runningFoldIndexed;
- runningReduce, runningReduceIndexed.

^KT-41422 fixed