Reword EXPECT_AND_ACTUAL_IN_THE_SAME_MODULE compilation error
Review: https://jetbrains.team/p/kt/reviews/9704

The warning message was unclear in terms of modules. The definition of
the module as the IJ module (Gradle project) is more popular and used
for the message's audience.

Alternative terms:
- "source set"
- "compilation unit"
- Something else?

I prefer "compilation unit" because it's more correct (there is no such
abstraction as "source set" on the level of the compiler) and robust
(what if tomorrow a new build system will be invented with its own terms
that clash with "source set"?)

See KT-55177 comments