[Gradle] ProjectMetadataProvider: Allow nullable compiled metadata for SourceSet
The commit leading to the issue described in KT-58319 was:
510eca9da1353f3c132160f1fe632c5c5f47b394
[Gradle] Implement source set inspection methods as 'suspend'

In this commit one method called
'getCommonSourceSetsForMetadataCompilation' was implemented
in a suspend style.

This re-implementation however introduced a semantic change to the function:
Previously this method (counterintuitively) returned commonMain
even if the SourceSet only participated in a single platform compilation.
This is because the KotlinMetadataTarget still cas a 'main' compilation
that includes 'commonMain'. The previous implementation did not filter
this compilation therefore found two platform types for the SourceSet,
assuming that it can be compiled to metadata.

This lead to a 'commonMain' compilation created which then gets
correctly disabled by 'isMetadataCompilationSupported'.

After the commit, a single target project did not create a 'commonMain'
compilation anymore. This leads to no metadata being provided
by the metadata provider. This seems OK, but the 'ProjectMetadataProviderImpl'
threw an error in cases the IDE requested this metadata.

^KT-58319 Verification Pending

(cherry picked from commit 8273328726592d3625ab4a79049117b5b4c0cba2)