[FIR] KT-58574: Ensure the incremental compilation works incorrectly
Note the test in
`IncrementalJsFirKlibCompilerWithScopeExpansionRunnerTestGenerated`.

This test is identical to
`serializedSubClassAndChangedInterfaces` except
the `A` and `B` classes have swapped their places.

Before being processed by the compiler, the input
files get sorted by their paths[1]. But if the class
`B` is compiled before the class `A`, then during
the generation of the data class members the call
to the property `a`'s `hashCode` becomes the call
to the original `Any::hashCode` function, not the
fake override `A::hashCode` (as it should). It
becomes a call to the fake override after the
incremental recompilation, though.

[1] `KotlinCoreEnvironment.kt:200`