[FIR] KT-58444: Prevent crash on consts without initializers
The crash happens because when the
IConstTransformer comes to the call to
`BATCH_SIZE`, it sees that it `canBeInterpreted`
(because it is `const`).
To interpret it properly, we must have its
initializer, but the presence of the initializer
depends on what `FirToConstantValueTransformer`
returns, and it may not properly transform some
constants (as is the case now with the `times`
function). So, there may be cases when some
`const` properties would have no initializers.

It looks like this is a separate more general
problem with how we evaluate constexprs, but for
now we may prevent the compiler crash with
an `if` check.

^KT-58444 Fixed