FIR/BI. Fix leakage of type variables from completed BI calls
Previously, builder inference finalization was performed after
joining CS of return arguments with call-tree
But, due to the fact that BI can postpone completion of arbitrary call
inside lambda (including return arguments) it was causing
calls that will be completed by BI to be added as non-completed return
arguments to the call-tree.
Then, after BI we were unable to find type variables that originated
from such calls during getOrderedAllTypeVariables as calls no longer
contain candidates

Instead of allowing to leak such calls as non-completed return arguments
we now perform builder inference completion before joining CS of
return arguments with the call-tree and effectively completing them to
a state similar to usual lambdas.
But, in order to construct an integration system for BI we need to
preliminary join CS of return arguments, and we perform it in temporary
CS

^KT-54294