[klib] Don't mess up order of decls when deserializing the whole klib
When the deserialization strategy is ALL (i.e. we deserialize
everything at once), preserve the order of top-level declarations
in which they were serialized (and exist in the source code).

Before this patch some declarations ("explicitly exported") were being
deserialized before everything else for reasons that this commit
explains: 7c4033f0bfd90d7767c101b3f7b9fd3bf2ed2515

This resulted in the incorrect order of declarations in
the deserialized IR. This didn't affect any behavior,
but fixing it is good for testing that the deserialized IR is the same
as the frontend-generated IR.