[Gradle] test lazy resolved configuration can handle circular dependencies
In gradle it is possible to add dependency to configuration which
is resolved to itself. It is possible due to nature of Configuration
that can contain both: dependencies and artifacts. When configuration
has dependency on itself, it practically means that this dependency
will be resolved to configuration's artifacts.

LazyResolvedConfiguration should be capable to handle such dependencies