Added a special case overload for CompositionLocalProvider
Added a special case for CompositionLocalProvider where a single
value is provided. This avoids overhead that makes providing
multiple values faster but is overhead for a signle value.

Bug: 288169379 ( https://issuetracker.google.com/issues/288169379 )
Test: ./gradlew :compose:r:r:tDUT
Relnote: """Added a speical case overload for CompositionLocalProviders
that avoids overhead used to make providing mulitple values faster
but is overhead when providing a single value."""

Change-Id: I6d640d97b96c26d9120c396063d84c73d947b852 ( https://android-review.git.corp.google.com/q/I6d640d97b96c26d9120c396063d84c73d947b852 )