Bump Kotlin/Native version in KGP to 1.9.20-dev-5069
Also, adjust NativeIrLinkerIssuesIT after 5f3dad1. See the explanation
below.

5f3dad1 makes -Xlazy-ir-for-caches disabled by default. So the compiler
now deserializes different subsets of IR and in different orders.
As a result, the error message for a test changes. Both error messages
(before and after this commit) are correct, and the selection depends
on the IR deserialization strategy.
Moreover, this commit changes the expected error message to the initial
version, used before -Xlazy-ir-for-caches was introduced.