KT-59801 [FIR] scripting: explicit import breaks highlighting
FIR-tree for scripts is built either for `KtScript` or `KtFile`
PSI node. The exact choice depends in particular on the existence of
explicit import statements. The latter leads to a lazy tree which is
completed on demand, for example for references shortening.

Tree completion (lazy bodies transformation) for scripts was not yet
supported. Hence, the error 'FirLazyExpression should be calculated
before accessing'.

This commit introduces the missing functionality.

^KT-59801 fixed