[Analysis API FIR] fix import reference resolve for unresolved import
To resolve the nested qualifier in the import statement
(see `FirReferenceResolveHelper.getSymbolsByResolvedImport`),
we create FirResolvedImport for the full import. Then, if it's resolved,
we create it for the import part we want to resolve and consider
that it's always resolved by casting it to `FirResolvedImport`.

It's possible that the outer import statement is resolved while the inner
one is not due to incorrect behavior of providers or project misconfiguration.

In the current case in the KT-59793, we find that `kotlinx.coroutines`
package exists while outer package `kotlinx` does not exist,
due to incorrect behaviour of `FirSymbolNamesProvider.getTopLevelCallableNamesInPackage`
which in IDE might return the whole set of packages which the project contain.

^KT-59793 fixed