Use RandomAccess sublist only if the AbstractList is RandomAccess
AbstractList.subList() would always return a sublist that is RandomAccess, even if the AbstractList is not a RandomAccess list, due to the private class SubList. This fixes that with the added extended class RandomAccessSubList. Now subList() will return a RandomAccess subList() if this is RandomAccess.

Fixes [KT-60207](https://youtrack.jetbrains.com/issue/KT-60207)