[Stubs] Move metadata path correctness check to the compiler
Normally, '.kotlin_builtins' files are thought to be loaded from the
Kotlin standard library JAR or some far JAR that bundles
the standard library.

However, in the 'r8.jar' from Android SDK, there are multiple
'.kotlin_builtins' files residing outside usual locations, in a nested
package. Such builtin definitions are not expected to be loaded.

The check was originally present in the IDE,
in 'readKotlinMetadataDefinition()'. However, stub indexing does not
go through that place.